+++ engine/flink/aws-managed-flink-deployment-manager/src/main/scala/pl/touk/nussknacker/engine/assembly/MergeStrategy.scala
@@ -3,6 +3,7 @@ package pl.touk.nussknacker.engine.assembly
 sealed trait MergeStrategy
 
 object MergeStrategy {
+  case object Concat              extends MergeStrategy
   case object Discard             extends MergeStrategy
   case object Deduplicate         extends MergeStrategy
   case object FilterDistinctLines extends MergeStrategy